organização e etc

+++ core/src/com/mygdx/game/logic/sprites/Bomb.java
@@ -64,7 +64,7 @@ public class Bomb extends Animated {
         TextureRegion temp = animations.get(0).getFrame(frame);
         //float Rotation = MathUtils.atan2(direction.y,direction.x)* MathUtils.radiansToDegrees;
 
-        batch.draw(temp,(float)(getWidth()/2  + getX() - temp.getRegionWidth()*0.5f),(float)(getHeight()/2  + getY() - temp.getRegionHeight()*0.5f),
+        batch.draw(temp,getWidth()/2  + getX() - temp.getRegionWidth()*0.5f,getHeight()/2  + getY() - temp.getRegionHeight()*0.5f,
                 temp.getRegionWidth()*0.5f, temp.getRegionHeight()*0.5f,
                 temp.getRegionWidth(), temp.getRegionHeight(),scale,scale,0);
